kernel: hil: rubble: clarify interfaces
